apps/api/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
sqlalchemy>=2.0.31
alembic>=1.13.2
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
redis>=5.0.7
celery>=5.4.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.9
pydantic-settings>=2.3.4
requests>=2.32.3
